Version 6.0.0
We have a brand new engine (Hit the red settings icon under the menu button to switch to it.) This will be enabled by default for new users and old users have the option to switch to it but will use the old engine by default. The old version will always be here but the new version is more secure. Closes #27

Added Freelancer preset.