Version 6.6.0
Removed Wish preset since it's full of scams
Added a colo(u)red underline to represent the typed password.
Fixed an issue with the theme selector (Closes #32)
Removed a "feature" I added in 0.4 where EACH CHARACTER TYPED acted as a new page. That was a bad idea.
Added presets for:
 - Square Enix
 - Trello